refactor: add grimoire background management functions and update imports for improved state handling
